+++ HISTORY.md
@@ -28,6 +28,8 @@
 * 2022-01-08
 	* Fixed: Class `ImplementationError`
 
+* 2022-01-15
+	* Added: more arguments for killProcesses()